day8

+++ src/cn/edu/zucc/personplan/comtrol/example/shoppingCartManager.java
@@ -1,10 +1,7 @@
 package cn.edu.zucc.personplan.comtrol.example;
 
 import cn.edu.zucc.personplan.itf.IShoppingCartManager;
-import cn.edu.zucc.personplan.model.BeanMerchant;
-import cn.edu.zucc.personplan.model.BeanOrderDetail;
-import cn.edu.zucc.personplan.model.BeanProductDetails;
-import cn.edu.zucc.personplan.model.BeanRider;
+import cn.edu.zucc.personplan.model.*;
 import cn.edu.zucc.personplan.util.BaseException;
 import cn.edu.zucc.personplan.util.BusinessException;
 import cn.edu.zucc.personplan.util.DBUtil2;
@@ -26,23 +23,25 @@ public class shoppingCartManager implements IShoppingCartManager {
             java.sql.PreparedStatement pst=conn.prepareStatement(sql);
             java.sql.ResultSet rs=pst.executeQuery();
             if(rs.next()) nowOrderId=rs.getInt(1)+1;
-            sql="select * from tbl_orderdetail where product_id=? and order_id=?";
+            sql="select * from tbl_orderdetail where product_id=? and order_id=? and user_id = ?";
             pst=conn.prepareStatement(sql);
             pst.setInt(1,productDetails.getProduct_id());
             pst.setInt(2,nowOrderId);
+            pst.setInt(3,BeanUser.currentLoginUser.getUser_id());
             rs=pst.executeQuery();
             if(rs.next()){
                 sql="update tbl_orderdetail set product_quantity=product_quantity+?,product_sumPrice=" +
-                        "product_sumPrice+?,product_discountPrice=product_discountPrice+? where order_id=? and product_id=?";
+                        "product_sumPrice+?,product_discountPrice=product_discountPrice+? where order_id=? and product_id=? and user_id = ?";
                 pst=conn.prepareStatement(sql);
                 pst.setInt(1,count);
                 pst.setFloat(2,productDetails.getProduct_price()*count);
                 pst.setFloat(3,productDetails.getProduct_discountPrice()*count);
                 pst.setInt(4,nowOrderId);
                 pst.setInt(5,productDetails.getProduct_id());
+                pst.setInt(6,BeanUser.currentLoginUser.getUser_id());
                 pst.execute();
             }else{
-                sql="insert into tbl_orderdetail values (?,?,?,?,?,?,?)";
+                sql="insert into tbl_orderdetail values (?,?,?,?,?,?,?,?)";
                 pst=conn.prepareStatement(sql);
                 pst.setInt(1,productDetails.getProduct_id());
                 pst.setString(2,productDetails.getProduct_name());
@@ -51,6 +50,7 @@ public class shoppingCartManager implements IShoppingCartManager {
                 pst.setInt(5,count);
                 pst.setFloat(6,productDetails.getProduct_price()*count);
                 pst.setFloat(7,productDetails.getProduct_discountPrice()*count);
+                pst.setFloat(8,BeanUser.currentLoginUser.getUser_id());
                 pst.execute();
             }
             rs.close();
@@ -74,10 +74,11 @@ public class shoppingCartManager implements IShoppingCartManager {
         Connection conn=null;
         try {
             conn= DBUtil2.getConnection();
-            String sql="delete from tbl_orderDetail where product_id = ? and order_id = ?";
+            String sql="delete from tbl_orderDetail where product_id = ? and order_id = ? and user_id=?";
             java.sql.PreparedStatement pst=conn.prepareStatement(sql);
             pst.setInt(1,orderDetail.getProduct_id());
             pst.setInt(2,orderDetail.getOrder_id());
+            pst.setInt(3,BeanUser.currentLoginUser.getUser_id());
             pst.execute();
             pst.close();
         } catch (SQLException e) {
@@ -100,9 +101,10 @@ public class shoppingCartManager implements IShoppingCartManager {
         Connection conn = null;
         try {
             conn= DBUtil2.getConnection();
-            String sql = "select * from tbl_orderDetail where merchant_id = ? and order_id=(select max(order_id) from tbl_orderDetail)";
+            String sql = "select * from tbl_orderDetail where merchant_id = ? and order_id=(select max(order_id)+1 from tbl_productOrder) and user_id = ?";
             java.sql.PreparedStatement pst = conn.prepareStatement(sql);
             pst.setInt(1,merchant.getMerchant_id());
+            pst.setInt(2, BeanUser.currentLoginUser.getUser_id());
             java.sql.ResultSet rs= pst.executeQuery();
             while(rs.next()){
               BeanOrderDetail bod = new BeanOrderDetail();
